+++ test/integration/property/propertiesCommon.js
@@ -169,7 +169,7 @@ function createDeclarations(targetClass, properties, options) {
             [itemName]: itemDeclaration
         };
 
-        const tenant = declaration[options.tenantName][options.applicationName];
+        const tenant = declaration[options.tenantName];
 
         properties.forEach((property) => {
             if (property.referenceObjects) {